finish implementing unary minus
adding unary operators to the pattern matching code was easier than i
thought! thanks zozi for the gud cood

the unary minus ended up being pretty intuitive to use, but there are
some things that are both weird to use and hard to fix. for instance, by
making it so -1^2 = -(1^2), it also makes it so -1 + 2 = -(1 + 2). i
already tried to fix that but it turned into a mess